SimPe 0.59a (QA) ---------------------------------------------------------     This is the first .NET 2.0 Release in the CVS. All Solutions are now located in the "__Solutions" Folder, the default build Target is the "__Release" Folder (even for Debug builds).

+++ fullsimpe/EnumComboBox/ExtProgressBar.cs
@@ -587,7 +587,7 @@ namespace Ambertation.Windows.Forms
 		{
 			if (tw==val) return;
 			tw = Math.Max(4, val);
-			tc =  Math.Max(2, (int)Math.Floor((Width-1) / (tw+2))); 	
+			tc =  Math.Max(2, (int)Math.Floor((double)((Width-1) / (tw+2)))); 	
 			CompleteRedraw();	
 			Invalidate();
 		}
